+++ R/module-define.R
@@ -594,7 +594,6 @@ setMethod(
 #'
 #' @keywords internal
 #' @importFrom data.table setnames
-#' @importFrom tools file_ext
 #' @rdname fillInputRows
 .fillInputRows <- function(inputDF, startTime) {
   factorCols <- sapply(inputDF, is.factor)
@@ -636,14 +635,14 @@ setMethod(
     if (any(is.na(inputDF2[, "fun"]))) {
       .fileExts <- .fileExtensions()
       fl <- inputDF2$file
-      exts <- na.omit(match(file_ext(fl), .fileExts[, "exts"]))
+      exts <- na.omit(match(fileExt(fl), .fileExts[, "exts"]))
       inputDF2$fun[is.na(inputDF2$fun)] <- .fileExts[exts, "fun"]
     }
 
     if (any(is.na(inputDF2[, "package"]))) {
       .fileExts <- .fileExtensions()
       fl <- inputDF2$file
-      exts <- match(file_ext(fl), .fileExts[, "exts"])
+      exts <- match(fileExt(fl), .fileExts[, "exts"])
       inputDF2$package[is.na(inputDF2$package)]  <- .fileExts[exts, "package"]
     }
     inputDF[!objectsOnly, ] <- inputDF2
@@ -658,7 +657,6 @@ setMethod(
 #'
 #' @keywords internal
 #' @importFrom data.table setnames
-#' @importFrom tools file_ext
 #' @rdname fillOutputRows
 .fillOutputRows <- function(outputDF, endTime) {
   needRenameArgs <- grepl(names(outputDF), pattern = "arg[s]?$")
@@ -690,7 +688,7 @@ setMethod(
   if (any(is.na(outputDF[, "fun"]))) {
     .fileExts <- .saveFileExtensions()
     fl <- outputDF$file
-    exts <- file_ext(fl)
+    exts <- fileExt(fl)
     if (any(is.na(fl)) | any(!nzchar(exts, keepNA = TRUE))) {
       outputDF$fun[is.na(fl) | (!nzchar(exts, keepNA = TRUE))] <- .fileExts$fun[1]
     }
@@ -703,12 +701,12 @@ setMethod(
   if (any(is.na(outputDF[, "package"]))) {
     .fileExts <- .saveFileExtensions()
     fl <- outputDF$file
-    exts <- file_ext(fl)
+    exts <- fileExt(fl)
     if (any(is.na(fl)) | any(!nzchar(exts, keepNA = TRUE))) {
       outputDF$package[is.na(fl) | (!nzchar(exts, keepNA = TRUE))] <- .fileExts$package[1]
     }
     if (any(is.na(outputDF[, "package"]))) {
-      exts <- file_ext(fl)
+      exts <- fileExt(fl)
       extsAvail <- checkKnownExts(exts, .fileExts)
       outputDF$package[is.na(outputDF$package)] <- .fileExts[extsAvail, "package"]
     }
@@ -740,3 +738,6 @@ addNamedEntry <- function(returnDataframe, templist, objectName, fn) {
   }
   returnDataframe
 }
+
+fileExt <- getFromNamespace("fileExt", "reproducible")
+filePathSansExt <- getFromNamespace("filePathSansExt", "reproducible")