+++ tests/testthat/test-simulation.R
@@ -270,8 +270,7 @@ test_that("simulation runs with simInit with duplicate modules named", {
   # laptop -- May 25, 2019 0.603 Seconds --> 120 microseconds/event!
   # laptop with new completed as environment -- May 25, 2019 0.357 Seconds --> 71 microseconds/event!
   options("spades.keepCompleted" = TRUE)
-  library(microbenchmark)
-  microbenchmark(times = nTimes, {spades(mySim, debug = FALSE)})
+  microbenchmark::microbenchmark(times = nTimes, {spades(mySim, debug = FALSE)})
 
   # Turn off completed list
   #  Changed to use "seconds" -- better comparison with simple loop
@@ -285,11 +284,11 @@ test_that("simulation runs with simInit with duplicate modules named", {
   # With many new "exists"
   # laptop -- May 25, 2019 0.264 Seconds --> 53 microseconds/event!
   options("spades.keepCompleted" = FALSE)
-  (a2 <- microbenchmark(times = nTimes, {spades(mySim, debug = FALSE)}))
+  (a2 <- microbenchmark::microbenchmark(times = nTimes, {spades(mySim, debug = FALSE)}))
   #profvis::profvis({for (i in 1:10) spades(mySim, debug = FALSE)})
 
   a <- 0
-  a3 <- microbenchmark(
+  a3 <- microbenchmark::microbenchmark(
     for (i in 1:N) {
       a <- a + 1
     }
@@ -308,12 +307,12 @@ test_that("simulation runs with simInit with duplicate modules named", {
   # Turn off completed list
   # New times using "second" -- Nov 26, 2018 0.443 Seconds --> 59 microseconds/event, even with sorting
   options("spades.keepCompleted" = FALSE)
-  (a2 <- microbenchmark(times = nTimes, {spades(mySim, debug = FALSE)}))
+  (a2 <- microbenchmark::microbenchmark(times = nTimes, {spades(mySim, debug = FALSE)}))
   #profvis::profvis({for (i in 1:10) spades(mySim, debug = FALSE)})
 
   # New times using "second" -- Nov 26, 2018 0.443 Seconds --> 130 microseconds/event, even with sorting
   options("spades.keepCompleted" = TRUE)
-  (a2 <- microbenchmark(times = nTimes, {spades(mySim, debug = FALSE)}))
+  (a2 <- microbenchmark::microbenchmark(times = nTimes, {spades(mySim, debug = FALSE)}))
 })
 
 test_that("conflicting function types", {