+++ R/zzz.R
@@ -42,9 +42,6 @@
                     env = pkgEnv
   )
 
-  ## import functions using backports:
-  backports::import(pkgname, "isFALSE")
-
   # parent <- parent.env(environment())
   # print(str(parent))
   #reg.finalizer(parent, finalize, onexit= TRUE)