+++ R/helpers.R
@@ -316,3 +316,7 @@ needInstall <- function(pkg = "methods", minVersion = NULL,
          "install.packages('",pkg,"')")
   }
 }
+
+isAbsolutePath <- getFromNamespace("isAbsolutePath", "reproducible")
+
+.isFALSE <- function(x) is.logical(x) && length(x) == 1L && !is.na(x) && !x