+++ tests/testthat/test-load.R
@@ -421,10 +421,12 @@ test_that("Filenames for simList", {
   s <- simInit()
   s$r <- raster(extent(0, 10, 0, 10), vals = 1, res = 1)
   s$r2 <- raster(extent(0, 10, 0, 10), vals = 1, res = 1)
-  s$r <- writeRaster(s$r, filename = tmpfile[1], overwrite = TRUE)
-  s$r2 <- writeRaster(s$r2, filename = tmpfile[3], overwrite = TRUE)
+  s$r <- suppressWarnings(writeRaster(s$r, filename = tmpfile[1], overwrite = TRUE))
+  s$r2 <- suppressWarnings(writeRaster(s$r2, filename = tmpfile[3], overwrite = TRUE))
   s$s <- stack(s$r, s$r2)
-  s$b <- writeRaster(s$s, filename = tmpfile[5], overwrite = TRUE)
+  a <- s$s
+  a[] <- a[]
+  s$b <- writeRaster(a, filename = tmpfile[5], overwrite = TRUE)
 
   Fns <- Filenames(s)