+++ tests/testthat/test-save.R
@@ -233,7 +233,6 @@ test_that("restart does not work correctly", {
   # Must be run manually
   setwd("~/GitHub/SpaDES.core")
   #devtools::install(update.dependencies = FALSE, dependencies = FALSE) # need to install latest so that at restart it has everything
-  devtools::load_all() # need the testInit function
   testInitOut <- testInit(libraries = "raster", tmpFileExt = c("grd", "Rdata", "Rdata"),
                           opts = list("spades.restartRInterval" = 1, "spades.moduleCodeChecks" = FALSE))
   on.exit({