+++ NEWS.md
@@ -8,7 +8,7 @@ version 1.0.4
 * none
 
 ## dependencies
-* completely removed `dplyr`, `lubridate`, `R.utils` from dependencies
+* completely removed `dplyr`, `lubridate`, `R.utils`, and `rlang` from dependencies
 * move `tcltk` to `Suggests`
 
 ## bug fixes